refactor: Refactored handlePlayerDelete in PlayerController and deletePlayer in PlayerEntityServiceImpl
1) It is now checked if the requested player exists, if not an exception is thrown
2) It is checked if the player is part of a lobby, if he is then he is removed from the lobby first before deletion
All others players are notified that the player leaves as in handlePlayerLeaveLobby
3) If the player is the only one in a lobby, the lobby and the player are deleted (leaveLobby is used for that)
All players are notified of the deleted Lobby as in handlePlayerLeaveLobby

In case of success Response Code RESPONSE_103 (player successfully deleted) is returned
In case of an exception Error Codes ERROR_2006 (player still exists), ERROR_2001 (player with the id does not exist) or exceptions from the leaveLobby Method are returned to the error queue

+++ src/main/java/at/aau/serg/websocketserver/service/impl/PlayerEntityServiceImpl.java
@@ -142,12 +142,19 @@ public class PlayerEntityServiceImpl implements PlayerEntityService {
 
     @Override
     public void deletePlayer(Long id) {
-        playerEntityRepository.deleteById(id);
+        Optional<PlayerEntity> playerEntityOptional = playerEntityRepository.findById(id);
+        if(playerEntityOptional.isPresent()) {
+            PlayerEntity playerEntity = playerEntityOptional.get();
+            GameLobbyEntity gameLobbyEntity = playerEntity.getGameLobbyEntity();
+
+            if(gameLobbyEntity != null) {
+                leaveLobby(playerEntity);
+            }
+            playerEntityRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException(ErrorCode.ERROR_2001.getErrorCode());
+        }
     }
-    // alternative suggestion:
-//    public void deletePlayer(PlayerEntity playerEntity) {
-//        playerEntityRepository.delete(playerEntity);
-//    }
